Billing: meter by actual bytes transferred, not rounded-up whole GB
A 2.5 GB file now consumes exactly 2.5 GB (was 3), and the wallet's GB
count falls fractionally as data moves. Buying is unchanged — whole-GB
packs in multiples of 10; only usage/metering becomes byte-accurate.

- migration 017: wallet_txns.units INTEGER -> NUMERIC(14,4) (fractional GB)
- billing.js: computeUnits = exact bytes/unitBytes (no ceil, no min-1);
  deductForUpload charges round(chargeableGB x rate); free first upload
  covers the first <=1 GB; unitsBalance/usedUnits now fractional
- api.js: /tenants used_units cast ::numeric
- billing.ejs + admin.ejs: fmtGb() formats fractional GB (balance, used, history)
- landing/faq/help/kb copy: "billed for actual data transferred — a 2.5 GB
  file is 2.5 GB, not rounded up"
- test: computeUnits asserts exact fractional GB (10/10 pass)

+++ src/web/routes/api.js
@@ -68,7 +68,7 @@ apiRouter.get('/tenants', requireSuperAdmin, wrap(async (_req, res) => {
     `SELECT t.id, t.slug, t.name, t.status, t.created_at,
             w.balance_paise, w.free_upload_used, w.unlimited,
             (SELECT COALESCE(SUM(wt.units), 0) FROM wallet_txns wt
-               WHERE wt.tenant_id = t.id AND wt.type = 'deduction')::int AS used_units,
+               WHERE wt.tenant_id = t.id AND wt.type = 'deduction')::numeric AS used_units,
             (SELECT u.email FROM users u WHERE u.tenant_id = t.id AND u.staff_permission IS NULL
                AND u.deleted_at IS NULL ORDER BY u.id LIMIT 1) AS owner_email,
             (SELECT count(*) FROM users u WHERE u.tenant_id = t.id AND u.deleted_at IS NULL)::int AS users,
